more changes from 'conf' to 'config' and renamed 'Util'
