+++ 2018 Competition Robot/src/org/usfirst/frc/team1732/robot/config/CTREConfig.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team1732.robot.conf;
+package org.usfirst.frc.team1732.robot.config;
 
 import com.ctre.phoenix.motorcontrol.NeutralMode;
 import com.ctre.phoenix.motorcontrol.VelocityMeasPeriod;
