+++ 2018 Competition Robot/src/org/usfirst/frc/team1732/robot/config/CTREParam.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team1732.robot.conf;
+package org.usfirst.frc.team1732.robot.config;
 
 public class CTREParam {
