+++ 2018 Competition Robot/src/org/usfirst/frc/team1732/robot/math/Utils.java
@@ -1,6 +1,6 @@
 package org.usfirst.frc.team1732.robot.math;
 
-public class Util {
+public class Utils {
 
 	public static double interpolate(double low, double high, double mu) {
 		return (high - low) * mu + low;
